qcacmn: Fix compilation issues caused by log format specifiers
Use proper format specifiers in dp_print_tso_stats,
also dp_tx_dump_tx_desc type cast variable as per
format specified to fix compilation issues.

CRs-Fixed: 3849167

+++ dp/wifi3.0/dp_tx.c
@@ -6567,7 +6567,7 @@ void dp_tx_dump_tx_desc(struct dp_tx_desc_s *tx_desc)
 		dp_tx_comp_warn("tx_desc->flags: 0x%x", tx_desc->flags);
 		dp_tx_comp_warn("tx_desc->id: %u", tx_desc->id);
 		dp_tx_comp_warn("tx_desc->dma_addr: 0x%x",
-				tx_desc->dma_addr);
+				(unsigned int)tx_desc->dma_addr);
 		dp_tx_comp_warn("tx_desc->vdev_id: %u",
 				tx_desc->vdev_id);
 		dp_tx_comp_warn("tx_desc->tx_status: %u",
